// src/screens/opRefundApproval/types.ts
type OpRefundApprovalItem = {
  RefundId: number;
  TokenNo: string;
  BranchId: number;
  PatientId: number;
  VisitId: number;
  UHID: "GS//00000001";
  PatientName: string;
  Age: string;
  Gender: string;
  TotalBillAmount: number;
  TotalDiscountPerOnBill: number;
  TotalDiscountAmountOnBill: number;
  RoundOff: number;
  TotalRefundAmount: number;
  IsRefundApproved: number;
  IsRefundCollected: number;
  ApprovalRemarks: string | null;
  IsLevel1Approve: number | null;
  Level1ApproveId: number | null;
  Level1ApproveOn: string | null;
  IsLevel2Approve: number | null;
  Level2ApproveId: number | null;
  Level2ApproveOn: string | null;
  IsLevel3Approve: number | null;
  Level3ApproveId: number | null;
  Level3ApproveOn: string | null;
  IsLevel4Approve: number | null;
  Level4ApproveId: number | null;
  Level4ApproveOn: string | null;
  IsCancel: number;
  Status: "Refund Approval Pending";
  StatusId: number;
  CancelBy: string | null;
  CancelOn: string | null;
  CancelReason: string | null;
  CreatedBy: string;
  CreatedOn: string;
  LastModifiedBy: string | null;
  LastModifiedOn: string | null;
  FlagId: number;
  CanApprove: number;
  RefundApprovedID: number;
  RefundApprovedName: string;
  RefundReason: string;
  RefundRemark: string;
};

type OpRefundApprovalCancelItem = {
  RefundId: number;
  TokenNo: string;
  BranchId: number;
  PatientId: number;
  VisitId: number;
  UHID: string;
  PatientName: string;
  Age: string;
  Gender: string;
  TotalBillAmount: number;
  TotalDiscountAmountOnBill: number;
  TotalDiscountPerOnBill: number;
  TotalRefundAmount: number;
  StatusId: number;
  Status: string;
  IsRefundApproved: number;
  IsRefundCollected: number;
  ApprovalFlowId: number | null;
  ApprovalFlow: string | null;
  IsAllApprovalRequired: number | null;
  ApprovalLevelId: number | null;
  ApprovalLevel: string | null;
  Level1ApproverNames: string | null;
  Level2ApproverNames: string | null;
  Level3ApproverNames: string | null;
  Level4ApproverNames: string | null;
  IsLevel1Approve: number | null;
  Level1ApprovedByName: string | null;
  Level1ApproveOn: string | null;
  IsLevel2Approve: number | null;
  Level2ApprovedByName: string | null;
  Level2ApproveOn: string | null;
  IsLevel3Approve: number | null;
  Level3ApprovedByName: string | null;
  Level3ApproveOn: string | null;
  IsLevel4Approve: number | null;
  Level4ApprovedByName: string | null;
  Level4ApproveOn: string | null;
  NextApprovalName: string | null;
  NextApprovalLevel: number | null;
  ApprovalRemarks: string | null;
  RefundApprovedID: number | null;
  RefundApprovedName: string | null;
  RefundReason: string | null;
  RefundRemark: string | null;
};

export type { OpRefundApprovalCancelItem, OpRefundApprovalItem };